+++ app/Http/Controllers/Api/SchemaDataStoreController.php
@@ -22,7 +22,7 @@ class SchemaDataStoreController extends BaseApiController
         );
 
         if ($validator->fails()) {
-            return $this->error($validator->errors(), 'Input validation error', $request->all(), 422);
+            return $this->error($validator->errors(), 'Input validation error', $request->all(), Response::HTTP_UNPROCESSABLE_ENTITY);
         }
 
         $schemaDataStore = new SchemaDataStore;
@@ -36,6 +36,9 @@ class SchemaDataStoreController extends BaseApiController
         }
 
         $data = isset($id) ? $schemaDataStore->where('id', $id)->first() : $schemaDataStore->get();
+        if((isset($id) && !isset($data)) || (!isset($id) && count($data) == 0)){
+            return $this->success([], 'No Items record(s) found', [], Response::HTTP_NOT_FOUND);
+        }
 
         return $this->success($data, 'Items successfully retrieved', [], Response::HTTP_OK);
     }
@@ -53,7 +56,7 @@ class SchemaDataStoreController extends BaseApiController
         );
 
         if ($validator->fails()) {
-            return $this->error($validator->errors(), 'Input validation error', $request->all(), 422);
+            return $this->error($validator->errors(), 'Input validation error', $request->all(), Response::HTTP_UNPROCESSABLE_ENTITY);
         }
 
         try {
@@ -71,9 +74,9 @@ class SchemaDataStoreController extends BaseApiController
                 throw new \RuntimeException('Could not save process category');
             }
 
-            return $this->success(['table' => $schema->name], 'Schema data successfully stored.', $request->all(), 200);
+            return $this->success(['table' => $schema->name], 'Schema data successfully stored.', $request->all(), Response::HTTP_CREATED);
         } catch (\Throwable $exception) {
-            return $this->error($exception->getMessage(), 'An error occurred while trying to store schema data.', []);
+            return $this->error($exception->getMessage(), 'An error occurred while trying to store schema data.', [], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
 
@@ -90,11 +93,14 @@ class SchemaDataStoreController extends BaseApiController
         );
 
         if ($validator->fails()) {
-            return $this->error($validator->errors(), 'Input validation error', $request->all(), 422);
+            return $this->error($validator->errors(), 'Input validation error', $request->all(), Response::HTTP_UNPROCESSABLE_ENTITY);
         }
 
         try {
             $schema = CRMSchema::find($request_data['schema_id']);
+            if((!isset($schema))){
+                return $this->success([], 'No schema record found to update', [], Response::HTTP_NOT_FOUND);
+            }
 
             $schemaDataStore = new SchemaDataStore;
             $schemaDataStore->getTable();
@@ -107,12 +113,12 @@ class SchemaDataStoreController extends BaseApiController
             }
 
             if ($schemaDataStore->save() === false) {
-                throw new \RuntimeException('Could not save process category');
+                throw new \RuntimeException('Could not save schema data');
             }
 
-            return $this->success(['table' => $schema->name], 'Schema data successfully stored.', $request->all(), 200);
+            return $this->success(['table' => $schema->name], 'Schema data successfully stored.', $request->all(), Response::HTTP_OK);
         } catch (\Throwable $exception) {
-            return $this->error($exception->getMessage(), 'An error occurred while trying to store schema data.', []);
+            return $this->error($exception->getMessage(), 'An error occurred while trying to store schema data.', [], Response::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
 
@@ -128,16 +134,22 @@ class SchemaDataStoreController extends BaseApiController
         );
 
         if ($validator->fails()) {
-            return $this->error($validator->errors(), 'Input validation error', $request->all(), 422);
+            return $this->error($validator->errors(), 'Input validation error', $request->all(), Response::HTTP_UNPROCESSABLE_ENTITY);
         }
 
         try {
             $schema = CRMSchema::find($request_data['schema_id']);
+            if((!isset($schema))){
+                return $this->success([], 'No schema record found to delete', [], Response::HTTP_NOT_FOUND);
+            }
 
             $schemaDataStore = new SchemaDataStore;
             $schemaDataStore->getTable();
             $schemaDataStore->setTable($schema->name);
             $schemaDataStore = $schemaDataStore->where('id',$id)->first();
+            if((!isset($schemaDataStore))){
+                return $this->success([], 'No schema data record found to delete', [], Response::HTTP_NOT_FOUND);
+            }
 
             if ($schemaDataStore->delete() === false) {
                 throw new \RuntimeException('Could not delete the item');
